app/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuro-preprocess"
version = "0.1.0"
description = "Neuroimaging preprocessing pipeline with OpenTelemetry instrumentation"
authors = [
    {name = "OpenTelemetry Learning Project"}
]
requires-python = ">=3.10"
keywords = ["neuroimaging", "preprocessing", "opentelemetry", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "numpy>=1.24.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# Phase 4: Will add OpenTelemetry dependencies
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
]

[project.scripts]
neuro-preprocess = "neuro_preprocess.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
neuro_preprocess = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
